Fix bug - ResourcesBar class

+++ src/main/java/ogame/ResourcesBar.java
@@ -13,31 +13,26 @@ public class ResourcesBar {
     public static final String DARKMATTER = "//*[@id=\"resources_darkmatter\"]";
     public static final String ENERGY_BALANCE = "//*[@id=\"resources_energy\"]";
 
-    public static int metal(WebDriver w) {
+    public static long metal(WebDriver w) {
         return value(w,METAL,"Doesn't download value of metal.");
     }
-
-    public static int crystal(WebDriver w) {
+    public static long crystal(WebDriver w) {
         return value(w,CRYSTAL,"Doesn't download value of crystal.");
     }
-
-    public static int deuterium(WebDriver w) {
+    public static long deuterium(WebDriver w) {
         return value(w,DEUTERIUM,"Doesn't download value of deuterium.");
     }
-
-    public static int darkmatter(WebDriver w) {
+    public static long darkmatter(WebDriver w) {
         return value(w,DARKMATTER,"Doesn't download value of darkmatter.");
     }
-
-    public static int energyBalanace(WebDriver w) {
+    public static long energyBalanace(WebDriver w) {
         return value(w,ENERGY_BALANCE,"Doesn't download value of energy.");
     }
-
-    private static int value(WebDriver w, String xPath, String errorMsg){
+    private static long value(WebDriver w, String xPath, String errorMsg){
         try {
             WebElement e = w.findElement(By.xpath(xPath));
             String value = e.getText().replace(".", "");
-            return Integer.parseInt(value);
+            return Long.parseLong(value);
         } catch (Exception ex) {
             AppLog.printOnConsole(ResourcesBar.class.getName(), 1, errorMsg);
         }
